projects/batch-jobs/pyproject.toml
[tool.poetry]
name = "batch-jobs"
version = "0.1.0"
description = ""
authors = ["bgasparotto <bruno.m.gasparotto@gmail.com>"]
readme = "README.md"
packages = [{include = "batch_jobs"}]

[tool.poetry.dependencies]
python = "^3.10"
pyspark = "^3.4.0"
psycopg2-binary = "^2.9.6"
pytest = "^7.3.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
